refactor: separate library and action source files
