+++ .gitattributes
@@ -1,4 +1,2 @@
-dist/** -diff linguist-generated
-restore/dist/** -diff linguist-generated
-save/dist/** -diff linguist-generated
+dist/action/** -diff linguist-generated
 pnpm-lock.yaml -diff linguist-generated
