+++ .prettierignore
@@ -1,4 +1,2 @@
 dist
-restore/dist
-save/dist
 pnpm-lock.yaml
