+++ lefthook.yml
@@ -36,7 +36,7 @@ pre-commit:
     - name: build action
       run: pnpm rollup -c
       glob:
-        - dist/*
+        - dist/action/*
         - src/*.ts
         - .npmrc
         - pnpm-lock.yaml
@@ -45,4 +45,4 @@ pre-commit:
         - src/*.test.ts
 
     - name: check diff
-      run: git diff --exit-code dist {staged_files}
+      run: git diff --exit-code dist/action {staged_files}
