+++ rollup.config.js
@@ -2,14 +2,14 @@ import { nodeResolve } from "@rollup/plugin-node-resolve";
 import typescript from "@rollup/plugin-typescript";
 
 export default [
-  { input: "src/main.ts", dir: "dist" },
-  { input: "src/post.ts", dir: "dist" },
-  { input: "src/restore/main.ts", dir: "dist/restore" },
-  { input: "src/save/main.ts", dir: "dist/save" },
-].map(({ input, dir }) => ({
+  "src/action/main.ts",
+  "src/action/post.ts",
+  "src/action/restore.ts",
+  "src/action/save.ts",
+].map((input) => ({
   input,
   output: {
-    dir,
+    dir: "dist/action",
     entryFileNames: "[name].bundle.mjs",
   },
   plugins: [nodeResolve(), typescript()],
