+++ src/lib/internal/cmd.test.ts
