+++ src/lib/internal/cmd.ts
