+++ tsconfig.build.json
@@ -1,9 +1,9 @@
 {
   "extends": "@tsconfig/node20",
-  "include": ["src/lib.ts"],
+  "include": ["src/lib"],
   "exclude": ["**/*.test.ts"],
   "compilerOptions": {
     "declaration": true,
-    "outDir": "dist"
+    "outDir": "dist/lib"
   }
 }
